refactor: separa logica de viwer ajuda
